pyproject.toml
[tool.poetry]
name = "jupyter-project"
version = "0.1.0"
description = "Jupyter notebook environment"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.0.0"
openai = "^1.53.0"
anthropic = "^0.38.0"
fpdf = "^1.7.2"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
numpy = "^2.1.3"
neo4j = "^5.26.0"
requests = "^2.32.3"
sentence-transformers = "^3.2.1"
swarm = {git = "ssh://git@github.com/openai/swarm.git"}
pydantic = "^2.9.2"
shortuuid = "^1.0.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
